Only populate defaults for input properties in MakeTerraformInputs
When creating terraform inputs, we would previously inject defaults for properties marked as being output-only (!Required and !Optional).

In particular, this caused AutoName transforms used in many of our TF-based providers to attempt to populate a default Name input even when the TF schema indicated Name was not an input property - leading to failures from the TF provider.

Fixes pulumi/pulumi-aws#250